requirements.txt
anthropic>=0.40.0
praw>=7.8.0
yfinance>=0.2.40
feedparser>=6.0.11
beautifulsoup4>=4.12.0
requests>=2.32.0
jinja2>=3.1.4
APScheduler>=3.10.4
pytz>=2024.1
python-dotenv>=1.0.1
lxml>=5.2.0
